requirements.txt
Django==4.2.6
djangorestframework==3.14.0
psycopg2==2.9.10
PyJWT==2.9.0
pyotp==2.9.0
qrcode[pil]==8.0
sendgrid==3.6.5
sendgrid-django==4.2.0
django-cryptography==1.1
django-q==1.3.9